+++ src/components/Views/consts.js
@@ -1,3 +1,3 @@
-export const VIEW_HIDDEN = 0;
-export const VIEW_BEFORE_HIDDEN = 2;
-export const VIEW_VISIBLE = 1;
+export const VIEW_HIDDEN = 'VIEW_HIDDEN';
+export const VIEW_BEFORE_HIDDEN = 'VIEW_BEFORE_HIDDEN';
+export const VIEW_VISIBLE = 'VIEW_VISIBLE';