+++ src/components/Views/types.js
@@ -2,10 +2,12 @@ import PropTypes from 'prop-types';
 
 const ViewTypes = {
   children: PropTypes.node.isRequired,
-  views: PropTypes.arrayOf(
+  config: PropTypes.arrayOf(
     PropTypes.shape({
       name: PropTypes.string.isRequired,
       component: PropTypes.elementType.isRequired,
+      path: PropTypes.string,
+      default: PropTypes.boolean,
     }),
   ).isRequired,
 };